Fixed a bug with missed reminders
The check for a missed reminder had a flaw in the logic. Originally, it would check if today date is greater than or equal to today AND if the current time was greater than or equal (+ 60 seconds grace period). This doesn't work if its a new day past the reminder date but it's still before the reminder time.
Now, it will check if today's date is greater than the reminder time OR if the current time is past the reminder time.