Fixed a bug with missed reminders / Fixed a bug with reminder name
Fixed the logic again for the missed reminders. It will check if the date has passed the reminder date OR if the date is the same but the time has passed (+ 60 seconds).
Fixed the reminder's date component of the reminder's name. The date will be the selected date and not today's date. This was occurring internally in task list (It was not affecting the checkbox list text)