Added more options to Modify / Added checks for playing sound, modifying, and removing / Reset audio text when custom audio is disabled
It is now possible to modify the custom audio of a reminder (If there wasn't one set, you can choose to set one).
Implemented try/catch checks to playing audio, removing a reminder entry, and modifying a reminder entry.
The textbox that stores the audio file path is reset when the custom audio option is disabled.